+++ backend/DEPENDENCIES.md
@@ -1,18 +1,7 @@
 # Backend Project Dependency List
-
-The following dependencies are required for the backend project and will be installed automatically with `npm install`:
-
 - @neondatabase/serverless
 - cors
 - dotenv
 - express
 - morgan
 - nodemon
-
-To install all dependencies, run the following command in the `backend` directory:
-
-```
-npm install
-```
-
-This will download and install all required packages as specified in `package.json`.
